Move all the tests into gym/gym/tests, this might fix all the import errors... hopefully
