+++ gym/tests/envs/robotics/hand/test_manipulate.py
@@ -3,7 +3,7 @@ import pickle
 import pytest
 
 from gym import envs
-from tests.envs.spec_list import skip_mujoco, SKIP_MUJOCO_WARNING_MESSAGE
+from gym.tests.envs.spec_list import skip_mujoco, SKIP_MUJOCO_WARNING_MESSAGE
 
 
 ENVIRONMENT_IDS = (
