+++ gym/tests/envs/robotics/hand/test_reach.py
@@ -3,7 +3,7 @@ import pickle
 import pytest
 
 from gym import envs
-from tests.envs.spec_list import skip_mujoco, SKIP_MUJOCO_WARNING_MESSAGE
+from gym.tests.envs.spec_list import skip_mujoco, SKIP_MUJOCO_WARNING_MESSAGE
 
 
 @pytest.mark.skipif(skip_mujoco, reason=SKIP_MUJOCO_WARNING_MESSAGE)
