+++ gym/tests/envs/test_registration.py
