+++ gym/tests/wrappers/test_resize_observation.py
